Moved msp430-libc and msp430mcu into the main tree
This is the beginning of a transition to get as much msp430 support into
the main tree as possible. Currently all that's left is getting a patch
for crossdev upstream (and a new version released) since the patches to
the various toolchain packages won't be put into the main tree until
they merge their efforts with the respective upstream projects.